Get Gmail Email Processing Pretty Done
This hooks up everything to register for webhooks for a User. It also
sets everything up to get the webhooks and convert these to the actual
messages we need to search through. Currently we are NOT actually
looking at the body of the emails, we are simply extracting it as a
Proof of Concept