feat: Add Bulma and add Navbar to Layout (#237)
In the process I installed Webpacker and got a JS pipeline going. This is how Bulma is running now, side by side with the asset pipeline for now.

Also upgraded to Rails 5.2 and got some specs running. They are just the ones that come with Clearance but that's gotta count for something right?

* Add webpacker and do webpacker install

* Get bulma added to the HTML

* Add navbar, it looks pretty reasonable. Static homepage is empty now

* Set some colors

* Install clearance views

* Ship some clearence feature specs. They fail right now because I don't require a password right now

* Get all specs passing by making password only optional if you have an omniauth authentication

* autocorrect

* Autogem rubocop todo

* Update all gems

* Lets just go to rails 5.2